changing existing file name from tab1.tsx an tab1.css to Profile.tsx and Profile.css
