--- src/pages/Tab1.css
@@ -1,68 +1,0 @@
-#present-alert, #open-action-sheet {
-    width: 95%;
-    margin: 5px auto;
-}
-
-#profile_pic {
-    width: 135px;
-    margin-top: -70px;
-    margin-bottom: -15px;
-    margin-left: 7px;
-}
-
-#modal_cover_pic {
-    width: 100%;
-    height: 100%;
-    margin: auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-#modal_prof_pic {
-    width: 100%;
-    height: 100%;
-    margin: auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-#coverpic{
-    /* width: 80%;
-    height: 75%; */
-    margin: auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-#profpic {
-    width: 80%;
-    height: 55%;
-    margin: auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-#modal_coverpic_close_button{
-    position: absolute;
-    margin-left: 360px;
-    margin-top: 70px;
-    /* background-color: transparent;
-    border-radius: 50px; */
-}
-
-#modal_profpic_close_button{
-    position: absolute;
-    margin-left: 360px;
-    margin-top: 70px;
-    /* background-color: transparent;
-    border-radius: 50px; */
-}
-
-.custom-modal {
-    --background: transparent;
-}
-
